+++ ghg_budget/artifact.py
@@ -1,10 +1,12 @@
 import pandas as pd
+from plotly.graph_objects import Figure
 from climatoology.base.artifact import (
     create_markdown_artifact,
     create_table_artifact,
     create_chart_artifact,
     _Artifact,
     Chart2dData,
+    create_plotly_chart_artifact,
 )
 from climatoology.base.computation import ComputationResources
 
@@ -127,19 +129,23 @@ def build_budget_comparison_chart_artifact(bar_chart_data: Chart2dData, resource
     )
 
 
-def build_time_chart_artifact(line_chart_data: Chart2dData, resources: ComputationResources) -> _Artifact:
-    return create_chart_artifact(
-        data=line_chart_data,
+def build_time_chart_artifact(line_chart: Figure, resources: ComputationResources) -> _Artifact:
+    return create_plotly_chart_artifact(
+        figure=line_chart,
         title='Entwicklung der CO₂-Emissionen in Heidelberg',
-        caption='Entwicklung der CO₂-Emissionen Heidelbergs seit 2016 (in 1000 Tonnen)',
-        description='Die Emissionswerte von 2016 bis 2021 sind Messwerte basierend auf dem BISKO-Standard, die Werte '
-        'ab 2022 sind Prognosen unter der Annahme, dass die zurzeit beschlossenen '
-        'Maßnahmen Heidelbergs zur Emissionsreduzierung umgesetzt werden. \n\nAnmerkung: Die Emissionswerte bilden '
-        'nicht die gesamten Emissionen der Stadt Heidelberg ab, sondern nur etwa 64 % der Emissionen. Dies liegt '
-        'daran, dass die Emissionen nach dem BISKO-Standard ermittelt wurden. Der BISKO-Standard erfasst vereinfacht '
-        'gesagt nur die Emissionen, die von Heidelberger:innen in Heidelberg ausgestoßen werden, nicht aber jene, die '
-        'beispielsweise von Heidelberger:innen außerhalb Heidelbergs verursacht werden. Mehr Informationen zur '
-        'BISKO-Systematik finden Sie links im Reiter "Berechnung des CO₂-Budgets".',
+        caption='Entwicklung der CO₂-Emissionen Heidelbergs und alternative Reduktionspfade seit 2016 (in 1000 Tonnen)',
+        description='Die Emissionswerte von 2016 bis 2021 in der Datenreihe "Prognose" sind Messwerte basierend auf '
+        'dem BISKO-Standard, die Werte ab 2022 sind Prognosen unter der Annahme, dass die zurzeit '
+        'beschlossenen Maßnahmen Heidelbergs zur Emissionsreduzierung umgesetzt werden. In dieser Prognose '
+        'wird Heidelberg erst 2068 CO₂-neutral. Die alternativen '
+        'Reduktionspfade stellen fiktive Szenarien dar, wie Heidelberg bis 2040 CO₂-Neutralität erreichen '
+        'könnte bzw. hätte erreichen können, ohne die CO₂-Budgets für die Einhaltung der Temperaturziele '
+        'zu überschreiten.\n\nAnmerkung: Die Emissionswerte bilden nicht die gesamten Emissionen der Stadt '
+        'Heidelberg ab, sondern nur etwa 64 % der Emissionen. Dies liegt daran, dass die Emissionen nach '
+        'dem BISKO-Standard ermittelt wurden. Der BISKO-Standard erfasst vereinfacht gesagt nur die '
+        'Emissionen, die von Heidelberger:innen in Heidelberg ausgestoßen werden, nicht aber jene, die '
+        'beispielsweise von Heidelberger:innen außerhalb Heidelbergs verursacht werden. Mehr Informationen '
+        'zur BISKO-Systematik finden Sie links im Reiter "Berechnung des CO₂-Budgets".',
         resources=resources,
         filename='time_chart',
         primary=True,